+++ elfo-core/src/message.rs
@@ -9,13 +9,53 @@ use smallbox::{smallbox, SmallBox};
 use crate::dumping;
 
 pub trait Message: fmt::Debug + Clone + Any + Send + Serialize + for<'de> Deserialize<'de> {
+    #[inline(always)]
+    fn name(&self) -> &'static str {
+        self._vtable().name
+    }
+
+    #[inline(always)]
+    fn protocol(&self) -> &'static str {
+        self._vtable().protocol
+    }
+
     #[doc(hidden)]
-    const VTABLE: &'static MessageVTable;
+    #[inline(always)]
+    fn labels(&self) -> &'static [Label] {
+        self._vtable().labels
+    }
+
+    #[doc(hidden)]
+    #[inline(always)]
+    fn dumping_allowed(&self) -> bool {
+        self._vtable().dumping_allowed
+    }
+
+    #[doc(hidden)]
+    #[inline(always)]
+    fn upcast(self) -> AnyMessage {
+        self._touch();
+        AnyMessage {
+            vtable: self._vtable(),
+            data: smallbox!(self),
+        }
+    }
+
+    // Private API.
+
+    #[doc(hidden)]
+    fn _vtable(&self) -> &'static MessageVTable;
 
     // Called while upcasting/downcasting to avoid
     // [rust#47384](https://github.com/rust-lang/rust/issues/47384).
     #[doc(hidden)]
     fn _touch(&self);
+
+    #[doc(hidden)]
+    #[inline(always)]
+    fn _erase(&self) -> dumping::ErasedMessage {
+        smallbox!(self.clone())
+    }
 }
 
 pub trait Request: Message {
@@ -24,6 +64,8 @@ pub trait Request: Message {
     #[doc(hidden)]
     type Wrapper: Message + Into<Self::Response> + From<Self::Response>;
 }
+
+// === AnyMessage ===
 
 // Reexported in `elfo::_priv`.
 pub struct AnyMessage {
@@ -35,36 +77,6 @@ pub struct AnyMessage {
 }
 
 impl AnyMessage {
-    #[inline]
-    pub fn new<M: Message>(message: M) -> Self {
-        message._touch();
-
-        AnyMessage {
-            vtable: M::VTABLE,
-            data: smallbox!(message),
-        }
-    }
-
-    #[inline]
-    pub fn name(&self) -> &'static str {
-        self.vtable.name
-    }
-
-    #[inline]
-    pub fn protocol(&self) -> &'static str {
-        self.vtable.protocol
-    }
-
-    #[inline]
-    pub fn labels(&self) -> &'static [Label] {
-        self.vtable.labels
-    }
-
-    #[inline]
-    pub fn dumping_allowed(&self) -> bool {
-        self.vtable.dumping_allowed
-    }
-
     #[inline]
     pub fn is<M: Message>(&self) -> bool {
         self.data.is::<M>()
@@ -94,17 +106,14 @@ impl AnyMessage {
         Ok(message)
     }
 
-    #[inline]
-    pub fn erase(&self) -> dumping::ErasedMessage {
-        (self.vtable.erase)(self)
-    }
-
+    #[doc(hidden)]
     #[cfg(feature = "network")]
     #[inline]
     pub fn write_msgpack(&self, buffer: &mut [u8]) -> Result<usize, rmp_serde::encode::Error> {
         (self.vtable.write_msgpack)(self, buffer)
     }
 
+    #[doc(hidden)]
     #[cfg(feature = "network")]
     #[inline]
     pub fn read_msgpack(
@@ -115,6 +124,27 @@ impl AnyMessage {
         lookup_vtable(protocol, name)
             .map(|vtable| (vtable.read_msgpack)(buffer))
             .transpose()
+    }
+}
+
+impl Message for AnyMessage {
+    #[inline(always)]
+    fn upcast(self) -> AnyMessage {
+        self
+    }
+
+    #[inline(always)]
+    fn _vtable(&self) -> &'static MessageVTable {
+        self.vtable
+    }
+
+    #[inline(always)]
+    fn _touch(&self) {}
+
+    #[doc(hidden)]
+    #[inline(always)]
+    fn _erase(&self) -> dumping::ErasedMessage {
+        (self.vtable.erase)(self)
     }
 }
 
@@ -131,9 +161,31 @@ impl fmt::Debug for AnyMessage {
     }
 }
 
-// Message Virtual Table.
+impl Serialize for AnyMessage {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::ser::Serializer,
+    {
+        // TODO: avoid allocation here
+        self._erase().serialize(serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for AnyMessage {
+    fn deserialize<D>(_deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::de::Deserializer<'de>,
+    {
+        Err(serde::de::Error::custom(
+            "AnyMessage cannot be deserialized",
+        ))
+    }
+}
+
+// === MessageVTable ===
 
 // Reexported in `elfo::_priv`.
+/// Message Virtual Table.
 #[derive(Clone)]
 pub struct MessageVTable {
     /// Just a message's name.