+++ elfo-test/src/proxy.rs
@@ -50,8 +50,9 @@ impl Proxy {
     pub fn send<M: Message>(&self, message: M) -> impl Future<Output = ()> + '_ {
         let location = Location::caller();
         self.scope.clone().within(async move {
+            let name = message.name();
             if let Err(err) = self.context.send(message).await {
-                panic!("cannot send {} ({}) at {}", M::VTABLE.name, err, location);
+                panic!("cannot send {} ({}) at {}", name, err, location);
             }
         })
     }
@@ -70,8 +71,9 @@ impl Proxy {
     ) -> impl Future<Output = ()> + '_ {
         let location = Location::caller();
         self.scope.clone().within(async move {
+            let name = message.name();
             if let Err(err) = self.context.send_to(recipient, message).await {
-                panic!("cannot send {} ({}) at {}", M::VTABLE.name, err, location);
+                panic!("cannot send {} ({}) at {}", name, err, location);
             }
         })
     }
@@ -80,9 +82,10 @@ impl Proxy {
     pub fn request<R: Request>(&self, request: R) -> impl Future<Output = R::Response> + '_ {
         let location = Location::caller();
         self.scope.clone().within(async move {
+            let name = request.name();
             match self.context.request(request).resolve().await {
                 Ok(response) => response,
-                Err(err) => panic!("cannot send {} ({}) at {}", R::VTABLE.name, err, location),
+                Err(err) => panic!("cannot send {} ({}) at {}", name, err, location),
             }
         })
     }