+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 
+## [Unreleased]
+### Changed
+- message: now `AnyMessage` implements `Message`, but it's still hidden.
+- message: add `Message::{name,protocol,labels,upcast}`.
+
 ## [0.2.0-alpha.1] - 2023-06-01
 ### Changed
 - **BREAKING** supervisor: actors discard `ValidateConfig` by default ([#49], [#87]).